requirements.txt
docx==0.2.4
docxcompose==1.4.0
keyboard==0.13.5
pyautogui==0.9.54
python-pptx
